feat: YOLO 프레임 처리 기능 추가 및 비디오 스트리밍 개선
- VideoProcessingService에 YOLO 프레임 생성 메서드 추가
- lives.py에서 프레임 생성 로직을 VideoProcessingService로 이동하여 코드 중복 제거
- 비디오 스트리밍 시 YOLO 검증 후 마킹된 프레임을 반환하도록 수정
